build: clean baseline warnings and wire intended design hooks
The warning baseline added in #16 was emitting ~340 warnings across
project sources (down from ~340 to ~70 after the previous tick_module.h
+ FetchContent SYSTEM cleanup). This commit drives the count to zero so
the gate can be flipped to -Werror.

Most edits are mechanical: unnamed parameters in default-noop virtuals
and stub implementations, static_cast at value-narrowing sites,
explicit empty {} for ActorTechnologyState aggregate inits,
[[maybe_unused]] on benchmark anti-DCE volatiles, and removal of dead
locals (abs_lat, community alias, drugs_id, ticks, amount_per_token,
expected_health, old_stage). Several real-design integrations were
also implemented or preserved rather than deleted:

* WorldState.Province.avg_property_value (new field) +
  RegionDelta.avg_property_value_update + apply_deltas wiring +
  RealEstateModule monthly write. The spec
  (simulation/modules/real_estate/real_estate_types.h §33.1 and
  docs/interfaces) calls for a per-province avg_property_value
  recomputed monthly; the module was computing it locally and
  discarding the result. The field is left out of the persistence
  schema for now (it's a UI/observability cache, recomputed every
  monthly tick) so no SCHEMA_VERSION bump is required.

* SettlementGenerator.create_businesses now bails out when
  owner_pool is empty (previously fell into rng.next_uint(0) -> 0
  followed by indexing an empty vector — UB only reachable from
  configs with npc_count == 0).

* DrugEconomyModule per-tier output_quality threading into
  BusinessDelta.output_quality_update (already shipped in 48ddb79;
  -Wunused-but-set-variable on output_quality was the original tell).

Scaffolding helpers (find_npc_delta, find_consequence,
make_test_player, lod1_default_transit_ticks, formal_output) are
restored with [[maybe_unused]] and explicit comments describing what
they will be wired into rather than silently deleted.

WorldGenerator.assign_archetype: province_count parameter is
documented unused but kept on the signature for the future
"diversity within first N provinces" rebalance hinted at by the
existing comment block.

Test scaffolding: world_generator_test now asserts has_criminal
(generator should always produce at least one criminal business);
the resources test no longer tracks has_sand since the comment
already explained why sand can be absent on landlocked highlands.

Determinism suite + 1313 ctest cases all pass.

Refs #16.

+++ simulation/cli/cli_main.cpp
@@ -221,8 +221,10 @@ static void print_metrics(const WorldState& world) {
                              : total_biz_cash / static_cast<float>(world.npc_businesses.size());
 
     std::printf("%-6u | %10.1f | %10.2f | %10.3f | %10.3f | %10.3f | %10.3f | %10.1f\n",
-                world.current_tick, avg_capital, avg_price, stability, crime, grievance, cohesion,
-                avg_biz_cash);
+                world.current_tick, static_cast<double>(avg_capital),
+                static_cast<double>(avg_price), static_cast<double>(stability),
+                static_cast<double>(crime), static_cast<double>(grievance),
+                static_cast<double>(cohesion), static_cast<double>(avg_biz_cash));
 }
 
 static bool check_nan_contamination(const WorldState& world) {
@@ -475,7 +477,8 @@ int main(int argc, char* argv[]) {
     for (const auto& p : world.provinces) {
         std::printf("  [%u] %-20s pop=%u infra=%.2f agri=%.2f deposits=%zu markets=%zu npcs=%zu\n",
                     p.id, p.fictional_name.c_str(), p.demographics.total_population,
-                    p.infrastructure_rating, p.agricultural_productivity, p.deposits.size(),
+                    static_cast<double>(p.infrastructure_rating),
+                    static_cast<double>(p.agricultural_productivity), p.deposits.size(),
                     p.market_ids.size(), p.significant_npc_ids.size());
     }
